app/requirements.txt
alembic==1.4.3
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
branca==0.4.1
bs4==0.0.1
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cite==0.6.0
click==7.1.2
coverage==5.3
cssselect2==0.4.1
cycler==0.10.0
defusedxml==0.6.0
dnspython==2.0.0
dominate==2.5.2
elasticsearch==7.9.1
email-validator==1.1.1
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-Cors==3.0.9
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-LoginManager==1.1.6
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
flask-statistics==1.0.2
Flask-Testing==0.8.0
Flask-WeasyPrint==0.6
Flask-WTF==0.14.3
folium==0.11.0
gunicorn==20.0.4
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.3
pandas==1.1.4
parinya==0.0.27
pdfkit==0.6.1
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
Pyphen==0.10.0
python-bcrypt==0.3.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-slugify==4.0.1
pytz==2020.1
requests==2.24.0
schedule==0.6.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
text-unidecode==1.3
tinycss2==1.1.0
Unidecode==1.1.1
urllib3==1.25.10
visitor==0.1.3
WeasyPrint==52.1
webencodings==0.5.1
Werkzeug==1.0.1
wkhtmltopdf==0.2
WTForms==2.3.3
